Move display_errors wrapping from logic to widget
The display_errors decorator wraps a function in try-catch and displays
exceptions in an error dialog.
I haven't run into an issue yet but I foresee that if a function wrapped
by dislay_errors ends up showing up more than once in a single call
stack that has an exception occuring in it, then the error message can
get displayed multiple times. So display_errors should only go on
functions that we expect will not show up more than once in a call
stack. This is difficult to keep up if display_errors is used on logic
functions, but for the most part easy to do if display_errors is used on
the widget-related functions. This is also where error dialogs are more
reasonable -- as a reaction to some user interaction rather than being
raised directly by low level code.

This is part of work towards #58 where the call stacks have been made
complicated enough to warrant the change.