Fix incorrect handling of transducer coordinates
Issue #58

The problem was noticed when loading an M2 transducer json file, which
was in units of 'm', and having the transducer look 1000x smaller that
it was supposed to. The transform we want to apply to the model node
should assume the model node geometry is given in *openlifu* coordinates
to begin with. Then the transform should convert

location (in openlifu coords) -->
transformed location according to openlifu transform (in openlifu coords) -->
transformed location according to openlifu transform (in slicer coords).

In this bugfix, the first arrow is transducer.matrix
and the second arrow is openlifu2slicer_matrix.

Previously I was converting the transducer *transform* itself to slicer
coordinates, which is not the goal when the object geometry is given in
openlifu coordinates.