Fix database path line edit file browser (#94)
The file dialog on the ctk lineedit was looking for files not directories.
Regression was introduced in b717911addd0e4d2bb597694de79557bc4cd1253
The reason it was introduced is that <set> on that filters property was
not appreciated by qt designer, which then dropped the property when we
saved the UI file. Setting to <string> instead seems to also work
and make qt designer happy.