Make convenience function to access data module parameter node
Part of work towards #77. I will need to access the loaded objects
from inside the SlicerOpenLIFUSession class, and I don't want to
have to define this class inside the data module. Will be nice
to not be too repetitive with the pattern for accessing the loaded
objects in the data module parameter node.