Load protocol alongside session (#109)
This also makes protocol a "session affiliated" object just like
transducer, volume, etc. So when a protocol is there as part of a
session, it gets cleaned up when the session is unloaded, just like
transducer. And when loading a session is going to replace a manually
loaded protocol, then there is a warning.