Write solution approvals/revokes to database (#113)
Solutions are meant to be write-once, with the exception of approval.
There's no need to complicate things by adding the concept of "save
solution", just for writting approval.
So here approval should get written to the database directly.

When a solution is by itself with no active session, then there's no
need to do anything with the databse.