Implement wip protocol cache and reloading (#169)
When working on protocols, if changes are unsaved, they are cached. The
cache is searched before the loaded protocols when loading protocols
with a given ID so that users can resume progress. Saving the protocol
and deleting the protocol will remove it from the cache. In the future,
reverting changes should do the same thing. Furthermore, it would be
nice to know in the protocol selector whether a protocol has unsaved
changes.