Add combo box wip notification (#169)
Closes #169

Added a "[ + ]" indicator to the name of the protocol in the combo box
when leaving an unfinished protocol. There were a few ways of monitoring
whether that indicator should be added, and in this solution we add the current
protocol text to the state of the logic. While this now requires keeping
track of additional clutter, it seemed to be the solution that added the
fewest headaches.