Implement login button update logic (#172)
Implement simple logic for displaying the login button and automatically
switching to the Home module if there is no admin in the database. This
was manually tested and works, but the default database has an admin, so
this functionality won't be observed much unless a dedicated database
with no admin role is created.

- Note that password hashes in the database aren't real hashes, so there
  is currently no way of logging in to any accounts. Suggested change is
  replacing the fake hashes with real hashes.

- There is no current logic to actually verify login tries. That should
  be done in `onLoginClicked`.