Finish login module prototype (#172)
A login module is now implemented that stores user state and allows
logins based on the DB (except that the db itself must be locally
updated to the latest version if testing). Without login, an anonymous
role is used. Without an admin in the database, admin is automatically
set. None of the roles are enforced, and new roles cannot be created.

- Added a convenience function to get the current openlifu user in
  OpenLIFULib. This will be equivalent to allowing any module to get the
  current user name, roles, etc. if so desired.