Implement selector marking of WIP protocols (#169)
Protocols that are cached WIPs (new and revertible) are marked with a
`[  +  ]` before the protocol name and id in the selector. Most of the
added logic handles adding and removing that string, but noticeably,
there was a second _is_ flag added, which is enabled and disabled
while the display is being updated from a protocol. This is because
trigger_unsaved_changes will cause a state change, which is erroneous
when we are programmatically changing the display to match a protocol.
The display is changed for a variety of reasons; for instance, every
single time `onProtocolSelectorIndexChanged()` is called the display
is programmatically changed, which happens every time
`reloadProtocols()` is called (as `reloadProtocols` will clear the
combo box and refill it from loaded and new protocols). With the
display being programmatically changed, we shouldn't be invoking the
signal that changes the save state to to SaveState.UNSAVED_CHANGES.

Why signals are blocked with a flag:

There were a few attempted approaches for blocking signals when the
display is updated. However, doing blockSignals(True) for every single
widget would require reconnection them after blockSignals(False). The
flag approach is much more concise.