Fix VF nodes loaded via session not getting watched (#145)
Virtual fit transform nodes need to be watched for modification so
that approval can be revoked if they are ever modified.
Virtual fit nodes were already getting watched when they are added via
the pre-planning module ("running" a virtual fit), but they were
not getting watched when they were loaded in from a session.