Replace clashing VF results on session load (#145)
This covers a rare situation where the user has some virtual fit results
that are affiliated with a session, and then invalidates the session and
doesn't clean up the virtual fit results. Later if they reload the
session, this now allows the dangling virtual fit results to get
replaced rather than giving an error about the conflict.