Implement send solution button using mock (#156)
Implements the function that sets the loaded_solution as the solution on
the device through the LIFUInterface. Currently, this uses
`test_mode=True`, which is the mock implementation. Furthermore, the
status is checked for readiness using the `get_status()` function. If
any exception is thrown, the send is considered failed. This function
uses try/except because it is currently assumed that device failures
will throw exceptions, but this may change.