Fix run enabledness without approved solution (#156)
When the solution is sent to the hardware, the button should NOT be
turned on directly, but instead the control flow for deciding whether
the button is enabled should be run. Furthermore, because solution
approval is more critical than the solution being sent, the control flow
was reordered so that approval is checked first.