Implement button to manually get device status (#140)
The button currently operates by instantiating a new LIFUInterface
object in test mode. It may be that, in the future, we need to hold
state to a single initialized object, which would be at an earlier point
in interacting with this module. Right now, the status is always
`LIFUInterfaceStatus.STATUS_READY` because this is all the mock outputs.