Integrate photocollection with database and session (#150)
This commit makes a lot of changes to the data module such that photo
collections stored in the temporary location are transferred into the
OpenLIFU-python database. Lots of the functions used for photoscans were
copied, but instead replaced with "photocollections", and because there
is no actual OpenLIFU-python photocollection object, the
photocollections associated with the session (and loaded in memory) are
referenced given their reference_number. reference_number is taken from
the terminology used in 3D Open Water. Because there is no wrapping of
OpenLIFU-python objects, some of the functionality is simpler, but right
now, after loading a photocollection in the data module, the files are
copied to the database.