Implement create new account button (#173)
A button to create a new account has been implemented which launches a
form dialog to input user data. Right now, only admins can access it,
but the button should be deactivated when a database is not connected as
well (it is not deactivated). There should also probably be some slight
restrictions on inputs for the form data, i.e. limits to how long the
name can be, what the user id can look like, etc.

- Tested and confirmed that a user is added into the data table with
  hash of salted password