Implement account management buttons (#173)
Implement the 3 buttons for interacting with user accounts in the
account management table. This involved creating a new dialog for
changing the password, connecting with the dialog for creating new user
accounts, and deleting accounts from the database. Note that there are
still some cases not figured out, such as a user deleting their own
account.