Separate connecting LIFUInterface into button (#140)
This commit adds a dedicated button for initializing the LIFUInterface,
along with a checkbox that allows test_mode to be toggled in the user
interface. When the LIFUInterface is connected, the logic holds its
state, and it can be reconnected at any time. This commit also adds
logic to button enabledness to require an initialized LIFUInterface as
well as reorganizes some of the code around that.