Integrate hardware signals with run progress (#140)
The progress of the run is reported through signals sent from the
device, which can be connected to a callback function which parses said
string. The current parsing routine is based on the following proposed
format by @georgevigelette and @peterhollender :

```
STATUS:RUNNING,MODE:SEQUENCE,PULSE_TRAIN:[1/1],PULSE:[5/100],TEMP_TX:32.23,TEMP_AMBIENT:28.9
STATUS:RUNNING,MODE:SEQUENCE,PULSE_TRAIN:[1/1],PULSE:[6/100],TEMP_TX:31.45,TEMP_AMBIENT:29.1
```

The current implementation parses this with regex and sets the
run_progress, which itself is used to update the GUI, as updates to that
property are listened for in the Widget setup. Note that the format
above may change, and there may be other things we want to parse from
the status update. Furthermore, even in test_mode, there is a dummy
signal connected, as in test_mode, the tx and hv device are set in
demo_mode, which have mock signals (@georgevigelette , we may want to
improve the mock).