Make reconnecting LIFUInterface reset solution sent (#140)
When reconnecting a LIFUInterface, the solution state is set to NOT_SENT
and the buttons are updated accordingly. Note that it could be the case
that constructing a new LIFUInterface may not actually remove the
previous solution from the device, but it is good to assume that if the
user tries to reconnect, they should have to re-send the solution.