Display connection status of LIFUInterface (#140)
Add a QFrame which contains a QLabel displaying the connected state of
the LIFU device. Since we switched to using a singleton, the state label
is set based on whether both LIFU devices (TX;HV) are connected.
Furthermore, the signals from LIFUInterface related to its connectedness
state changing update the widget through the callbacks registered in the
logic.