Invoke start_monitoring in sonication control (#140)
This commit both creates an event loop that schedules the asynchronous
(and never-ending) task of `start_monitoring` on the LIFUInterface,
which will allow for USB connections to be discovered. While the
solution does not seem elegant, it does work, and for test_mode, a mock
of the `start_monitoring` was created that repeatedly calls a print
statement. When switching into `test_mode`, the task in the event loop
is switched.