Replace bf.Pulse form with abstract form widget (#165)
This commit sets the stage for abstractly inserting widgets that
represent OpenLIFU classes into placeholders. Because there may be
future changes in OpenLIFU-python, this solution minimizes the level of
coordination needed between the two codebases. It introduces a generic
framework that leverages Python class introspection to dynamically
create a QWidget that simply lists the fields of the class (dataclasses
work the best with this framework) and defines a simple API for
returning the class from the fields and setting the fields from an
instance of the class. The GUI ends up looking slightly programmatic and
raw, as it uses the member names for QLabel s, but this will be easily
solved by lightweight inheritance and renaming of the QLabels.