Replace bf.FocalPattern form with abstract widget (#165)
Similar for individual class forms, this commit sets the stage for
creating widgets that allow selection from a collection of
implementations of an abstract class and then editing of the fields in
that selection. The first example of this is with Wheel, but many other
constructs in openlifu have a protocol that can select from multiple
implementations of an ABC. This commit allows everything to be
dynamically populated, thus reducing repeated code and making things
easier to maintain if there are changes in openlifu object definitions,
or if more ABC implementations are made of, for example, a FocalPattern.