Support dynamic tuple types in widget creation (#165)
Extended the widget creation logic to handle tuples of arbitrary length
and type by dynamically generating child widgets based on the element
types. A `QWidget` container with a horizontal layout is used to group
the individual input widgets.

Each element in the tuple is mapped to a corresponding input widget
(`QSpinBox`, `QDoubleSpinBox`, `QLineEdit`, or `QCheckBox`) depending on
its type.

Tuples containing unsupported types, such as `dict`, are currently not
handled and will return `None`. That is, currently we cannot create a
form entry for a class that has a field that might look like this:

```python
@dataclass
class SimSetup(DictMixin):
    dims: Tuple[str, str, str]
    names: Tuple[str, dict, dict]  # <--- not allowed
    dict_names: dict  # <--- still allowed
```