Add parameter constraints form to protocol config (#165)
Note that, as a member of the Protocol class, parameter constraints are
not a particular object, but a dictionary. Therefore, we do not leverage
the abstract class definition form widget, and instead for parameter
constraints we replace with a singular DictTableWidget. This is
different from other cases in that in the other cases, the Protocol
member was typically an entire class that may have included a dict
within.