Prevent initialization of dataclass fields without init (#165)
Some @dataclass fields in OpenLIFU-python are not included in the
constructor for said class. For instance, `param_ids` below is not part
of `__init__` for `Material` (as `init=False`). Therefore, this commit
prevents widgets in the class form from being created if those fields
cannot be initialized. Furthermore, since this only applies to
@dataclass, `OpenLIFUAbstractClassDefinitionFormWidget` was renamed to
`OpenLIFUAbstractDataclassDefinitionFormWidget`

```
@dataclass
class Material:
\# ...
    param_ids: Annotated[Tuple[str, str, str, str, str], OpenLIFUFieldData("Parameter IDs", None)] = field(default_factory=lambda: ("sound_speed", "density", "attenuation", "specific_heat", "thermal_conductivity"), init=False, repr=False)
    """TODO: Add description"""
```