Move database_is_loaded from database parameter node to logic (#181)
Certain modules (OpenLIFULogin and OpenLIFUProtocolConfig) relied on the
database_is_loaded flag in the OpenLIFUDatabase parameter node to
determine whether features requiring a database are enabled. Parameter
node update signals were leveraged to react to those changes in
callbacks. The new mechanism uses the logic, and not the parameter node,
to hold the flag, and modules which react to database changes register
their callbacks by getting the OpenLIFUDatabase logic and adding their
callbacks to those called when the database_is_loaded property is
changed.