Add default database resources to Database module (#229)
Add the default database files to the Database module resources and
modify the utility to copying the default database. The default utility
will now create the database in the location if the User selects "Yes"
to creating the database in the location, which is prompted if a
database does not exist at that path. The QSettings are always set, even
if the user selects "No", as the path selected is independent from the
act of loading & initializing the database.