Make database init safe (#229)
Changes `def copy_preinitialized_database` so that instead of walking
over the destination and changing the file mode, it assumes that a
partial database might exist there, or that it might be a directory
containing files unrelated to openlifu. A database is initialized over
the area in conjunction with existing files if a database location is
chosen where there exists files. If a user chooses to initialize their
database in an in-use location, this may lead the user to overwrite
files on their filesystem.