Move ui element creation to before connections in data module (#229)
Opening OpenLIFUData when launching the app yielded the following error:

```
SlicerOpenLIFU-Debug % ./SlicerWithSlicerOpenLIFU
Switch to module:  "Welcome"
Populating font family aliases took 156 ms. Replace uses of missing font family "Courier" with one that exists to avoid this cost.
Switch to module:  "OpenLIFUData"
Traceback (most recent call last):
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUData.py", line 1261, in onNodeAdded
    self.updateLoadedObjectsView()
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUData.py", line 1053, in updateLoadedObjectsView
    self.loadedObjectsItemModel.removeRows(0,self.loadedObjectsItemModel.rowCount())
AttributeError: 'OpenLIFUDataWidget' object has no attribute 'loadedObjectsItemModel'
```

Which is presumably because onNodeAdded is called sometime after being
added as a callback in the data module, but before the ui elements
related to loadedObjectsItemModel are instantiated. This commit moves
the instantiation of UI elements to earlier in the widget setup, which
removes this error message.