Enforce guided mode visibility for static ui content (#128)
This commit creates a mechanism for hiding static content in the ui
file. Widgets can have a dynamic property called
`slicer.openlifu.hide-in-guided-mode` which, **when it exists at all as
a property for the widget, regardless of what the value is**, the widget
will be hidden when guided mode is activated. Some widgets in the ui
were given this property, and the Workflow object in the home module
will loop over all the modules, find the widgets that have that
property, and then hide them.

Currently, this does not deal with the case for more complex widgets
that may dynamically choose which widgets to hide. Such examples are the
OpenLIFUAlgorithmInputWidget, which dynamically populates widgets and
may have to make a decision of which widgets to hide. There is currently
a precedent set for this marked as TODO. The plan is to add those
widgets into a list and then call a function that the widget has to
define.