Enhance `OpenLIFUAlgorithmInputWidget` for guided mode (#128)
Modified the AlgorithmInput dataclass to include a reference to the
`QLabel` associated with each `QComboBox` in the
`OpenLIFUAlgorithmInputWidget`. Previously, the `QLabel` was implicit
and could not be referenced. However, now, the `QLabel` can be
referenced explicitly for each combo box.

This change enables programmatic control over label visibility alongside
their corresponding combo boxes. The `enforceGuidedModeVisibility()`
method was added for the `OpenLIFUAlgorithmInputWidget` to hide both the
`QLabel` and `QComboBox` when guided mode is enforced (i.e. when
`enforceGuidedModeVisibility()` is called by the `Workflow` class.