Fully implement `Workflow.enforceGuidedModeVisibility()` (#128)
This finishes the implementation of `enforceGuidedModeVisibility()`.
Static widgets with the dynamic property are located and hidden in
guided mode, and widgets that wish to define a more complex version of
`enforceGuidedModeVisibility()` are added to the list, after which it is
called with the enforcement status. Currently, only
`OpenLIFUAlgorithmInputWidget` implements this, but there is an easily
extensible framework in place for adding new guided mode rules.