Hide/show user account banners based on user acct mode (#197)
This makes sure to update visibility of banners when user account mode
is activated. Note that the visibility of all banners is initialized to
False. This is because the widgets always exist (in application memory)
before user_account_mode is set. This is because all widgets are
dependencies of the login module.