Subclass dict[str, ParameterConstraint] definition in protocol config (#282)
The OpenLIFUParameterConstraintsWidget subclasses DictTableWidget to
create custom behavior for defining ParameterConstraint objects.
The param_constraints is a dictionary that maps parameter names to
constraints, and we want users to be able to intuitively add these
constraints. To achieve this, a custom dialog was made, based on the
dialog for adding abstract class values to the DictTableWidget rows, but
instead allowing a user to configure a parameter constraint.