Add utility CreateStringDialog (#262)
Add a widget that is a dialog simply for entering a string. It's the
analog to CreateKeyValueDialog but for lists. It is used by the
ListTableWidget when the item to enter into the list is not a generic
class type (e.g. TargetConstraint, for adding an arbitrary object to a
list) but is a string.