Remove GUI update and add additional comments (#308)
This remove the parameter node's update of the GUI after having its path
updated. This is because Slicer's dynamic property infrastructure will
keep the widget and parameter node in sync by a construct defined in the
`.ui` file called `SlicerParameterName`.

For instance, the widget's line edit will *automatically update to show
the latest setting of the parameter node* even if the parameter node was
the only thing updated, and no connection was made in `setup()` or in
the `onParameterNodeModified`:

```python
slicer.util.getModuleWidget('OpenLIFUDatabase')._parameterNode.databaseDirectory = Path('/')
```

The above snippet will update the widget GUI to also display the new
path as `/`.