Subclass SolutionAnalysisOptions definition in protocol config (#282)
* Also change replace_widget to deal with replacing the DictTableWidget
  in the form for defining SolutionAnalysisOptions

This commit subclasses SolutionAnalysisOptions to deal with the special
case of param_constraints needing the subclassed version of
DictTableWidget. When creating the form widget for defining
SolutionAnalysisOptions in the protocol config, it used the
DictTableWidget for defining the param_constraints field. However, a
special subclassed version of this was defined to give a more intuitive
way of defining `ParameterConstraint` objects.

The way that `DictTableWidget` is replaced with
`OpenLIFUParameterConstraintsWidget` is using `replace_widget`. However,
`replace_widget` was not made to deal with `QFormLayout`, only the
simple `QVBoxLayout`. Therefore, the widget replacement yielded the
following traceback:

```
Traceback (most recent call last):
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUProtocolConfig.py", line 269, in setup
    self.solution_analysis_options_definition_widget = OpenLIFUSolutionAnalysisOptionsDefinitionFormWidget()
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUProtocolConfig.py", line 2123, in __init__
    replace_widget(old_param_constraints_dicttablewidget, new_param_constraints_widget)
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFULib/util.py", line 187, in replace_widget
    layout.insertWidget(index, new_widget)
AttributeError: QFormLayout has no attribute named 'insertWidget'
```

This was fixed by adding a case where `replace_widget` handles the
parent of the widgets you are trying replace being a QFormLayout, so
that both the label and widget are replaced (a form has a label and a
widget for each row).