Remove caching of enforced permissions (#262)
This commit removes caching of enforced permissions. The number of
widgets enforcing permissions is not enough to be noticeable on login
state changes. Furthermore, it gets rid of a problem where if an admin
account's `user_id` is cached as having been enforced, and then they
remove the admin role from themself, they will still be considered an
admin user. This is because even though the login state *is* updated after
closing the console (through `updateWidgetLoginState`), the cache state
will result in no changes being made.

An alternative solution to the problem above is to store the cached
roles instead of `user_id`s, but it's easier to eliminate the cache
altogether.