Make changing active user reset loaded protocol data (#262)
Since certain users may have access to certain data through their role,
we unload all data when the user account changes in any way. Note that
we don't register callbacks in the Data module because everything has to
be done in the login module. Callbacks cannot be registered in the setup
for other widgets because other widgets are initialized earlier in the
dependency graph, due to the fact that the Login module depends on all
others.