Reset logged in user after user management panel edits (#262)
A user who is able to use the user management panel may end up editing
their own user in the database. Adding or removing roles needs to be
reflected, or any other changes to the current user, and so after the
console closes we match the user in the database and then update it, as
it could potentially have been updated.