Cache last active user when applying user changed updates (#262)
Cache the last user for which user updates were applied. The reason for
this rarely comes up, but what could happen is that a user could open up
the management console to modify other users, and then when closing the
console it would reset loaded data (e.g. protocols and sessions). While
it's a rare case, we don't want the user to be confused as to why their
data was unloaded just because they opened the management console.

However, on the contrary, if any edits are made to the currently
logged-in user, after opening the management console, all the data is
wiped as usual (as it tests for exact equality). We could make this less
strict and just do user-id equality, but the idea here was that if the
user changes we should probably assume it's a different user.