Change protocol database dialog to use full protocols (#262)
This changes the dialog for loading protocols from the database to use
complete `openlifu.plan.Protocol` objects. This is because the dialog
will need to conditionally display protocols available to load based on
the roles of the user.  Currently a user who cannot access a protocol
will be able to see the protocol in the database, but if they try to
load it, nothing happens. This behavior is unintuitive.