Prevent active_user from being None (#262)
In some cases, the active user was default_anonymous_user. When opening
the account management table and then closing it, a match could not be
found for default_anonymous_user in the database, and so the user was
set to None. This led to bugs and confusion, so this commit prevents the
active user from ever being None and fixes that bug.