Show all protocols when user account mode off (#262)
Fixes an issue where the active user was set to something (e.g.
default_anonymous_user) but user account mode was not on. The protocol
config would still filter out protocols based on what active user was
set to in the login logic, despite user account mode features supposed
to not be doing anything.