Subclass `SegmentationMethod` definition in protocol config (#318)
This commit subclasses the `SegmentationMethod` configurer to account
for the unique inheritance structure of `SegmentationMethods`. In brief,
there are three levels of inheritance, but the last two levels are
supposed to be alternative selections. Between the alternative
selections, two of them (`UniformWater` and `UniformTissue`) are
supposed to be the same as `UniformSegmentation` except that
`ref_material` is set as water/tissue and is not intended to be mutable.
The subclassing of the definition form takes all these details into
account.

[See the discussion on GitHub
here.](https://github.com/OpenwaterHealth/OpenLIFU-python/pull/289#discussion_r2073610129)