Make ABC implementations default construct when alternating (#318)
When configuring protocols, switching between ABC implementations (e.g.
for SegmentationMethod) will always default construct a new
`SegmentationMethod`. This is a subtle but important detail because it
means if a user edits fields under one implementation of a ABC and then
switches to a different implementation of the ABC, they will lose their
progress. In order to save all derived versions of the ABC, an entire
caching mechanism would be required to save temporary inputs to a class
that is orphaned from any protocol (i.e. it is a WIP alternate ABC
selection -- it is not actually defined as part of the protocol).