Increase default decimal precision for float definitions (#318)
In some rare cases, pulse durations past 1e-5 were specified in a
protocol, which was then set to 0.0000s in the protocol configuration.
This is because a default precision of 4 decimal places was set, and
there is no `__post_init__` defined on `Pulse` in openlifu that
enforces positivity of `Pulse.duration`. The following traceback
occurred when attempting to compute a sonication solution with a
`Pulse.duration` set to 0:

```python
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUSonicationPlanner.py", line 323, in onComputeSolutionClicked
    self.logic.computeSolution(activeData["Volume"], activeData["Target"],
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUSonicationPlanner.py", line 620, in computeSolution
    solution_openlifu, pnp_aggregated, intensity_aggregated, analysis_openlifu = compute_solution_openlifu(
  File "SlicerOpenLIFU-Debug/lib/Slicer-5.8/qt-scripted-modules/OpenLIFUSonicationPlanner.py", line 580, in compute_solution_openlifu
    solution, simulation_result_aggregated, scaled_solution_analysis = protocol.calc_solution(
  File "Slicer-Debug/python-install/lib/python3.9/site-packages/openlifu/plan/protocol.py", line 334, in calc_solution
    solution.scale(transducer, self.focal_pattern, analysis_options=analysis_options)
  File "Slicer-Debug/python-install/lib/python3.9/site-packages/openlifu/plan/solution.py", line 277, in scale
    analysis = self.analyze(transducer, options=analysis_options)
  File "Slicer-Debug/python-install/lib/python3.9/site-packages/openlifu/plan/solution.py", line 160, in analyze
    p0_Pa = np.max(output_signal, axis=1)
  File "Slicer-Debug/python-install/lib/python3.9/site-packages/numpy/core/fromnumeric.py", line 2810, in max
    return _wrapreduction(a, np.maximum, 'max', axis, None, out,
  File "Slicer-Debug/python-install/lib/python3.9/site-packages/numpy/core/fromnumeric.py", line 88, in _wrapreduction
    return ufunc.reduce(obj, axis, dtype, out, **passkwargs)
ValueError: zero-size array to reduction operation maximum which has no identity
```

An issue was opened up in
OpenLIFU-python for this
([OpenLIFU-python#302](https://github.com/OpenwaterHealth/OpenLIFU-python/issues/302)).

Setting the decimal precision to 8 does fix this case, but a decimal
precision of 8 may be adjusted when addressing the application
prettiness issue. Ideally, a more modest default precision should be
set, and then forms which need different levels of precision should
subclass and use the utility function `modify_widget_spinbox` to adjust.