Add protocol validity indicator (#334)
Assist the user configuring a protocol by displaying a small box showing
an exception that would be thrown if a protocol object was created with
the active parameters. The protocol is "checked" for exceptions each
time any field is updated, and the box is only showed when the editor is
enabled.

It may be helpful to note that the indicator is properly updated when
switching between erroneous and non-erroneous protocols (e.g. write
an erroneous protocol and get indicator as WIP -> switch to
non-erroneous protocol and see indicator go away -> switch back to
erroneous WIP and see indicator return) because changing the display to
match a protocol will call the `trigger_invalid_protocol_check` for each
label/box that was updated to match the contents of the erroneous
protocol, as those updates were connected to the signals emitted when
updates happen here:

```python
for f in [trigger_unsaved_changes, trigger_invalid_protocol_check]:
    self.ui.protocolNameLineEdit.textChanged.connect(f)
    self.ui.protocolIdLineEdit.textChanged.connect(f)
    self.ui.protocolDescriptionTextEdit.textChanged.connect(f)

    self.allowed_roles_widget.table.itemChanged.connect(lambda *_: f)
    self.pulse_definition_widget.add_value_changed_signals(f)
    self.sequence_definition_widget.add_value_changed_signals(f)
    self.abstract_focal_pattern_definition_widget.add_value_changed_signals(f)
    self.sim_setup_definition_widget.add_value_changed_signals(f)
    self.abstract_delay_method_definition_widget.add_value_changed_signals(f)
    self.abstract_apodization_method_definition_widget.add_value_changed_signals(f)
    self.segmentation_method_definition_widget.add_value_changed_signals(f)
    self.parameter_constraints_widget.table.itemChanged.connect(lambda *_: f)
    self.target_constraints_widget.table.itemChanged.connect(lambda *_: f)
    self.solution_analysis_options_definition_widget.add_value_changed_signals(f)
    self.virtual_fit_options_definition_widget.add_value_changed_signals(f)
```