Fix exact equality check on transforms matrices (#259)
Since there could be rounding or precision issues when the floats are
read in and out, they could match without being literally equal bit for
bit. So here it is better to use np.isclose which does a "fuzzy"
equality check based on a resonable default for relative and absolute tolerance.