Hide user account mode banner in guided mode (#344)
This adds the `slicer.openlifu.hide-in-guided-mode` property set to
`True` to all `UserAccountBanner` widgets at the end of their
construction so that they are hidden in guided mode. The current
intended guided mode experience is that the banner showing the logged-in
account and return button is hidden. This adds the property to
constructed widgets.

Note: There may be unintended consequences if
`enforceGuidedModeVisibility()` is called before the `UserAccountBanner`
widgets are replaced in the ui of each module. This should not occur, as
it is called after all `OpenLIFU.*Widget`s are setup() and ui elements
are replaced.