Refactor database and data module coupling (#231)
This commit finishes the logical coupling between database and data
modules, specifically around database loading. Database loading was
heavily tied to the subject/session selector due to the design of
certain functions. The responsibility of updating the selector in the
data module is now handled entirely in the data module, which does add a
callback to listen for changes in the database module's loaded database.