Fix adding photocollections and photoscans to loaded session (#369)
This fixes the photocollection capture, photocollection import, and add
photoscan button routines ("on.*Clicked") which previously were
implemented erroneously. The intention of these routines is to load data
to the currently loaded session, but the way they were implemented was
they looked at the currently selected session index in the previous
(removed) subject/session selector tree. This means that if a user
intended to add a photocollection to the current session but had a
different one selected, it would have added to that session but still
temporarily loaded the photocollection/photoscan because this behavior
was added to happen all the time.

Instead, the name of the functions was changed to better resemble their
behavior, and the photocollections/photoscans are added to the currently
loaded session.

Furthermore, in the new subject selector, when selecting sessions, the
photoscan/photocollection is added to the session that is selected in
real time, allowing a user to edit sessions without loading them, only
in the dialog.