Add loading subjects into data module logic (#369)
Subjects can now be loaded with a corresponding subject status viewer
via a load subjects button. It is added in conjunction with other
buttons for viewing subjects and loading sessions. The current subjects
is held in the logic, which, when changed, propagates updates to status
windows. If the subject is None, the volumes and sessions collapsible
buttons are disabled. Within the loading subjects window, you can add
subjects as well.

Note that loading subjects is entirely disconnected from the method of
loading sessions. Both functionalities exist freely and override each
other where values are set. For instance, if a subject is loaded, and
then a session for a different subject is loaded, that is possible. The
previous subject will be loaded in the logic, but the session loaded
will correspond to that one chosen in the subject/session selector.
Currently, loading subjects is purely an informational feature.