Overhaul sessions collapsible session (#369)
This commit fully deprecates the use of the subject/session selector
(and removes all functionality with error messages indicating such). All
the functionality is instead enabled in the Session collapsible section,
which allows sessions to be loaded for a given subject. Functions to
modify sessions were refactored, such as for adding photocollections and
photoscans. "Unload session" functionality was removed; when you load a
session you can load a different one or create a new one and load that.
These actions effectively unload the session.

Due to the fact that session-related matters are tightly coupled with
the parameter node, enabledness updates for session-level buttons were
added to the parameter node update routine. If a different subject is
loaded, the session is also unloaded. Functionality can be tweaked to
make this do nothing if the same subject is selected, but for now,
things unload when changes are made.