Mask the background out of loaded MRIs (#387)
Closes #387

One thing I slightly dislike about this method is that it modifies the
values in the volume array: it makes the background voxels take a value
that is 1 less than the minimum value in the volume. This way they can
simply be thresholded out. There should be other ways to do this that
involve creating a segmentation node out of the foreground mask and then
using it as a "cropping ROI" for the volume. This would be nicer but
then it's another piece of data to manage. This method  keeps things
simple, and it doesn't seem to harm the downstream skin segmentation
computation.

Another thing I slightly dislike about this method is that it takes
some time to compute the foreground mask, so this makes loading a volume
into the scene a slower process. But it isn't too bad, at least on my
system, and the result is definitely very pleasing to look at.