Automatically load chosen database location (#339)
When the dialog for choosing the database location is activated, the
user will automatically load a database when the location is chosen.
This required a custom implementation of ctkPathLineEdit browse()

Also change the name from "change database location" to "choose database
location". Reflects the new purpose of the button which is to load the
database.