Add valid directory indicator to database path (#339)
This indicator is relevant when editing the ctkPathLineEdit directly.
Each time a character is changed, the os module is used to verify that
the path currently in the text box is a directory or is not a directory.

Hitting return will load the directory. The color is not seen when the
path is out of focus.