Change database directory line edit coloring (#339)
When the path is modified but a database has not been connected, there
is a yellow border around the line edit. When a database is loaded
(`onDatabaseChanged`), the color is green. This color also applies when
loading a new a Database via the button. `updateParametersFromSettings`
also needed to be modified because updates to the line edit from
qsettings cause the box to turn yellow when the application starts up.