Add permissions-related tooltips to dynamic widgets (#369)
Note that only dynamic widgets have permissions-related tooltips when
disabled because the disable routine is managed by each individual
widget, where the tooltip context is obvious.

The current implementation of permissions across the app still happens
centrally in the Login module, accessing all permissions widgets across
the app, and so the rest of the application **does not** have
permissions-related tooltips.

Co-authored-by: Sadhana Ravikumar <sadhana.ravikumar@kitware.com>