Enable choosing number of images to use in mesh recon (#389)
Instead of specifying the image skip rate, you can specify the number of
images (roughly) to use. This is because in #389 we need to set
something like an "angular resolution" of photos that were captured in a
loop, and the image skip rate cannot be used to do this because the
total number of images in a photocollection can very.