Add sorting for subject/session tables (#403)
Note: Sorting must be turned off while appending rows to the table or
reloading the table. This is because row indexing does not work properly
while rows are sorted. This means that adding a subject or session will
usually render the table unsorted after the addition has been made, but
sorting can return as normal by clicking a column header.