Disable workflow save/exit based on session (#406)
The global `Workflow` class (holds reference to each modules workflow
controls) now holds reference to a loaded session. The Data parameter
update routine passes any updates to the loaded session onto the Home
module logic's reference to the global Workflow instance. When there is
no session, all of the save/exit buttons are disabled and vice versa
when there is a loaded session.